misc/data_quality_is/TasksManagement/requirements.txt
sqlalchemy=1.1.13
pymysql=0.7.11
celery=4.1
flower=0.9.2
pymongo=3.4
requests=2.18.4
python-dateutil=2.7.2